add routes, refactor App.js to remove clutter
