+++ frontend/src/pages/Dashboard.js
@@ -1,0 +1,5 @@
+const Dashboard = () => {
+  return <div>dashboard placeholder</div>;
+};
+
+export default Dashboard;
